+++ resources/views/blocks/ad-block.blade.php
@@ -44,13 +44,13 @@
                         @auth
                         @if(Auth::getUser()->hasLiked($ad))
                         <div class="spinner-border spinner-border-sm d-none" role="status"><span class="sr-only">Cargando...</span></div>
-                        <i class="lni-thumbs-down"></i>
+                        <i class="lni-heart-filled"></i>
                         @else
                         <div class="spinner-border spinner-border-sm d-none" role="status"><span class="sr-only">Cargando...</span></div>
-                        <i class="lni-thumbs-up"></i>
+                        <i class="lni-heart"></i>
                         @endif
                         @else
-                        <i class="lni-thumbs-up"></i>
+                        <i class="lni-heart"></i>
                         @endauth
                     </a>
                 </div>